backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
pydantic[email]==1.10.12
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
